[Refactor] TokenService 리펙트링
